+++ tests/ProxyInterfaceSourceGeneratorTests/Destination/Microsoft.SharePoint.Client.ClientObjectProxy.g.cs
@@ -33,25 +33,25 @@ namespace ProxyInterfaceSourceGeneratorTests.Source.PnP
 
         public object Tag { get => _Instance.Tag; set => _Instance.Tag = value; }
 
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         public global::Microsoft.SharePoint.Client.ObjectPath Path { get => _Instance.Path; }
 
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         public string ObjectVersion { get => _Instance.ObjectVersion; set => _Instance.ObjectVersion = value; }
 
-        [Microsoft.SharePoint.Client.PseudoRemoteAttribute]
+        [global::Microsoft.SharePoint.Client.PseudoRemoteAttribute]
         public bool? ServerObjectIsNull { get => _Instance.ServerObjectIsNull; }
 
         public global::ProxyInterfaceSourceGeneratorTests.Source.PnP.IClientObject TypedObject { get => MapToInterface(_Instance.TypedObject); }
 
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         public virtual void FromJson(global::Microsoft.SharePoint.Client.JsonReader reader)
         {
             global::Microsoft.SharePoint.Client.JsonReader reader_ = reader;
             _Instance.FromJson(reader_);
         }
 
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         public virtual bool CustomFromJson(global::Microsoft.SharePoint.Client.JsonReader reader)
         {
             global::Microsoft.SharePoint.Client.JsonReader reader_ = reader;
@@ -59,13 +59,13 @@ namespace ProxyInterfaceSourceGeneratorTests.Source.PnP
             return result__636829107;
         }
 
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         public void Retrieve()
         {
             _Instance.Retrieve();
         }
 
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         public void Retrieve(params string[] propertyNames)
         {
             string[] propertyNames_ = propertyNames;